Other amd64 warning fixes

+++ aes.c
@@ -58,7 +58,7 @@ compute_hash(const char* digest_name, const char* input, size_t input_len)
   const EVP_MD *md;
   EVP_MD_CTX *mdctx;
   unsigned char *md_value;
-  size_t md_len;
+  unsigned int md_len;
 
   OpenSSL_add_all_digests();
   md = EVP_get_digestbyname(digest_name);
